+++ src/main/java/qupath/ext/align/gui/interactive/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains the {@link qupath.ext.align.gui.interactive.AlignWindow AlignWindow} that provides
+ * ways to interactively align images, the {@link qupath.ext.align.gui.interactive.AlignCommand AlignCommand}
+ * that can start an align window and other classes used internally by the align window.
+ */
+package qupath.ext.align.gui.interactive;
