+++ src/main/java/qupath/ext/align/gui/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package contains the UI elements of the extension.
+ * This package contains the {@link qupath.ext.align.gui.AlignExtension AlignExtension} and some
+ * {@link qupath.ext.align.gui.Utils UI utilities}.
  */
 package qupath.ext.align.gui;
