+++ src/main/resources/qupath/ext/align/gui/interactive/image_selector.fxml
